holy shit that's a lot. tl;dr: command bar, lots of optimizations, text rendering, abstractions, a lotttttt

+++ src/engine/instances/datatypes/Color3.h
@@ -3,9 +3,9 @@
 
 namespace Nyanners::DataTypes {
   struct Color3 {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
-    uint8_t alpha;
+    int r;
+    int g;
+    int b;
+    int alpha;
   };
 }
